# Speed up RGB <--> XYZ conversions