This improves `srgb_compand` and `invert_srgb_compand`(renamed rgb->srgb) with almost no loss in accuracy. (#351)