@kimikage 2019-09-17
# Benchmark
```julia
julia> versioninfo()
Julia Version 1.2.0
Commit c6da87ff4b (2019-08-20 00:03 UTC)
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, skylake)
```
## before (be3e641d2193df646f1696cdce3d33d8dc62c7ee)
```julia
julia> @benchmark convert(RGB{Float64}, x) setup=(x=XYZ{Float64}(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     76.185 ns (0.00% GC)
  median time:      163.042 ns (0.00% GC)
  mean time:        170.986 ns (0.00% GC)
  maximum time:     435.670 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     970

julia> @benchmark convert(XYZ{Float64}, x) setup=(x=RGB{Float64}(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     5.382 ns (0.00% GC)
  median time:      202.278 ns (0.00% GC)
  mean time:        209.065 ns (0.00% GC)
  maximum time:     391.925 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     966

julia> @benchmark convert(XYZ{Float64}, x) setup=(x=RGB24(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     8.411 ns (0.00% GC)
  median time:      207.477 ns (0.00% GC)
  mean time:        214.632 ns (0.00% GC)
  maximum time:     480.270 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     963
```
## after
```julia
julia> @benchmark convert(RGB{Float64}, x) setup=(x=XYZ{Float64}(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     20.059 ns (0.00% GC)
  median time:      47.944 ns (0.00% GC)
  mean time:        50.982 ns (0.00% GC)
  maximum time:     145.436 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     997

julia> @benchmark convert(XYZ{Float64}, x) setup=(x=RGB{Float64}(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     5.416 ns (0.00% GC)
  median time:      54.764 ns (0.00% GC)
  mean time:        56.797 ns (0.00% GC)
  maximum time:     159.478 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     997

julia> @benchmark convert(XYZ{Float64}, x) setup=(x=RGB24(rand(),rand(),rand()))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     6.899 ns (0.00% GC)
  median time:      7.600 ns (0.00% GC)
  mean time:        7.897 ns (0.00% GC)
  maximum time:     46.901 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```