@kimikage 2019-09-19
# Assessment of the elimination of the two methods.
As [Codecov says](https://codecov.io/gh/JuliaGraphics/Colors.jl/src/be3e641d2193df646f1696cdce3d33d8dc62c7ee/src/conversions.jl), the current test-set does not have the test cases which call [the two methods](https://github.com/JuliaGraphics/Colors.jl/blob/be3e641d2193df646f1696cdce3d33d8dc62c7ee/src/conversions.jl#L170-L171).

The chains of conversion which may call the two methods should have at least one of the two type of (sub) chain (except the *direct* conversion).
- upstream `X` → `RGB24` → `AbstractRGB`
- `RGB24` → `AbstractRGB` → downstream `Y`

## Direct Conversion
The *direct* means that it is not intermediate conversion, i.e. it starts with `RGB24` and ends with `AbstractRGB`.
Let me get straight to the point, there is no *direct* conversion which call the two methods. The conversion completes within ColorTypes.jl as the following comment suggests:
https://github.com/JuliaGraphics/Colors.jl/blob/be3e641d2193df646f1696cdce3d33d8dc62c7ee/src/conversions.jl#L33-L35

- `convert(AbstractRGB, RGB24())`
  - of course, this calls [`convert(::Type{C}, c::C) where {C<:Colorant} = c`](https://github.com/JuliaGraphics/ColorTypes.jl/blob/7ef745605d6a0897078712a0a7a156c51295b559/src/conversions.jl#L6).(practically, inlined)
- `convert(AbstractRGB{N0f8}, RGB24())`
  - same as above. (`RGB24<:AbstractRGB{N0f8} == true`)
- `convert(RGB, RGB24())`
  - this calls [`ColorTypes._convert(RGB{N0f8}, RGB, RGB24, RGB())`](https://github.com/JuliaGraphics/ColorTypes.jl/blob/7ef745605d6a0897078712a0a7a156c51295b559/src/conversions.jl#L25).
- `convert(RGB{Float32}, RGB24())`
  - almost the same as above except the first argument (`RGB{Float32}`).

(to be continued)