@kimikage 2019-09-20
Since I found that the two methods are unused and have no effect on performance, I will, *paradoxically*, revert the elimination.
My original intention was that the two methods should be eliminated as a result of fixing the latter one as I commented in [[1]](https://github.com/JuliaGraphics/Colors.jl/pull/352#discussion_r325633412), [[2]](https://github.com/JuliaGraphics/Colors.jl/pull/352#discussion_r325411756), and [[3]](https://github.com/JuliaGraphics/Colors.jl/pull/352#discussion_r325642711).
Althoug the intention was based on the premise that they may be called, it was different to the reality. So, eliminating [the two](https://github.com/JuliaGraphics/Colors.jl/blob/be3e641d2193df646f1696cdce3d33d8dc62c7ee/src/conversions.jl#L170-L171) and [this](https://github.com/JuliaGraphics/Colors.jl/blob/be3e641d2193df646f1696cdce3d33d8dc62c7ee/src/conversions.jl#L389) are totally different issues. Thus, the issue should be report separately.

I will modify them and `push -f`. Then, let's merge it!