@timholy 2019-09-21
No worries, merge commits are fine as far as I'm concerned. I decide on a case-by-case basis whether I want to keep commits separate. For single-commit PRs I try to remember, but I forget frequently.

Thanks for tackling this!